fls

Struct ChannelSettings

pub(crate) struct ChannelSettings {
    pub(crate) input: DdsSettings,
    pub(crate) output: DdsSettings,
    pub(crate) amp: u24,
    pub(crate) lockin_k: Lockin<Lowpass<2>>,
    pub(crate) min_power: i32,
    pub(crate) clear: bool,
    pub(crate) phase_scale: [[i32; 2]; 2],
    pub(crate) iir: BiquadReprTree<Q32<29>, i32>,
    pub(crate) pow_gain: i16,
    pub(crate) hold_en: bool,
    pub(crate) iir_amp: BiquadReprTree<f32, f32>,
}

Fields
input: DdsSettings

Input (demodulation) DDS settings Feedback to stabilize the RF input phase is applied to the RF output on top of the output frequency and phase.

For the demodulation this is the total (DDS minus Lockin, i.e. lower sideband) demodulation frequency. If the modulation AOM is passed twice at +1 order, input/freq should be twice output/freq.

output: DdsSettings

Output (modulation) DDS settings

amp: u24

Demodulation amplitude control register.

Value

AD9959 amplitude control register (24 bits, see datasheet)

Default

0 for full scale amplitude and multiplier disable

lockin_k: Lockin<Lowpass<2>>

Lockin lowpass time constant. The lowpass is a cascade of one second order IIR filters, 12 dB/octave. This needs to be high enough to suppress the unwanted demodulation components and harmonics but as low as possible to maximize bandwidth. Many demodulation components and harmonics are also suppressed by the zeros of the batch size moving average FIR filter and judicious choice of lockin_freq.

TODO: settle pll and lockin settings into design after confirming optimal choice

Default

lockin_k = [0x200_0000, -0x2000_0000]

min_power: i32

Minimum demodulated signal power to enable feedback. Note that this is RMS and that the signal peak must not clip.

Value

log2 of the signal power relative to full scale. Range: [-63..0]

Default

min_power = -24 corresponding to about -69 dBFS.

clear: bool

Clear the phase unwrap tracking counters once. To make this setting edge-sensitive, after setting it to true, it must be reset to false by the user before setting any other settings.

phase_scale: [[i32; 2]; 2]

Scaling factor of the unwrapped phase and fine rational offset. The phase scaling is located after the phase unwrapping before the feedback IIR filter.

FIXME: doc rational offset

Value

[[phase_factor, phase_shr], [time_factor, time_shr]]

Default

phase_scale = [[1, 16], [0, 0]]: clamped range: ยฑ33 k turn (tracked range is ยฑ2 G turn) quantization: 1.5 ยต turn, 0.1 Hz

iir: BiquadReprTree<Q32<29>, i32>

Feedback IIR filter settings. The filter input is phase, the output is frequency.

Default

A proportional gain=-1 filter.

pow_gain: i16

Phase offset feedback gain. Phase feedback is a proportional bypass of the unwrapper, the IIR (including its input and output scaling) and the frequency feedback path. The phase offset gain is pow_gain/(1 << 13) rad/rad.

Value

Integer scaled phase feedback gain. Range: [-0x2000, 0x2000]

Default

0 for no phase feedback

hold_en: bool

Allow digital input to hold

iir_amp: BiquadReprTree<f32, f32>

Amplitude IIR filter. The filter input is squared magnitude, the output is DDS amplitude.

Default

No feedback

Implementations

impl ChannelSettings

pub(crate) fn update_phase_scale(&mut self)