fls

Struct Stream

#[repr(C)]
pub(crate) struct Stream {
    pub(crate) demod: Complex<i32>,
    pub(crate) phase: [i32; 2],
    pub(crate) delta_ftw: i32,
    pub(crate) delta_pow: i16,
    pub(crate) mod_amp: u16,
    pub(crate) pll: u32,
}

Stream data format.

Fields
demod: Complex<i32>

Demodulated signal. -1 << 31 corresponds to negative full scale.

phase: [i32; 2]

Current number of phase wraps. In units of turns.

delta_ftw: i32

Current frequency tuning word added to the configured modulation offset mod_freq.

delta_pow: i16

Current phase offset word applied to the modulation DDS.

mod_amp: u16

Modulation DDS amplitude word

pll: u32

PLL time